+++ OGLight/OGLight.meta.js
@@ -1,5 +1,5 @@
 // ==UserScript==
-// @version 4.8.1
+// @version 4.8.2
 // @name OGLight
 // @namespace    https://github.com/igoptx/ogameTools
 // ==/UserScript==